+++ app/tests/challenges_tests/test_tasks.py
@@ -16,7 +16,6 @@ from grandchallenge.challenges.tasks import (
     update_challenge_compute_costs,
     update_challenge_results_cache,
 )
-from grandchallenge.invoices.models import PaymentStatusChoices
 from grandchallenge.utilization.models import JobWarmPoolUtilization
 from grandchallenge.utilization.tasks import create_job_warm_pool_utilizations
 from tests.algorithms_tests.factories import AlgorithmJobFactory
@@ -71,155 +70,6 @@ def test_challenge_creation_from_request():
     assert challenge.short_name == challenge_request.short_name
     # requester is admin of challenge
     assert challenge_request.creator in challenge.admins_group.user_set.all()
-
-
-@pytest.mark.django_db
-def test_challenge_budget_alert_email(settings):
-    challenge = ChallengeFactory(short_name="test")
-    challenge_admin = UserFactory()
-    challenge.add_admin(challenge_admin)
-    staff_user = UserFactory(is_staff=True)
-    settings.MANAGERS = [(staff_user.last_name, staff_user.email)]
-    invoice = InvoiceFactory(
-        challenge=challenge,
-        support_costs_euros=0,
-        compute_costs_euros=10,
-        storage_costs_euros=0,
-        payment_status=PaymentStatusChoices.PAID,
-    )
-    phase = PhaseFactory(challenge=challenge)
-    evaluation = EvaluationFactory(
-        submission__phase=phase,
-        time_limit=60,
-    )
-
-    evaluation.utilization.invoice = invoice
-    evaluation.utilization.compute_cost_euro_millicents = 5 * 1000 * 100
-    evaluation.utilization.save()
-    update_challenge_compute_costs()
-
-    # Budget alert threshold not exceeded
-    assert len(mail.outbox) == 0
-
-    evaluation = EvaluationFactory(
-        submission__phase=phase,
-        time_limit=60,
-    )
-    evaluation.utilization.invoice = invoice
-    evaluation.utilization.compute_cost_euro_millicents = 3 * 1000 * 100
-    evaluation.utilization.save()
-    update_challenge_compute_costs()
-
-    # Budget alert threshold exceeded
-    assert len(mail.outbox) == 3
-    recipients = {r for m in mail.outbox for r in m.to}
-    assert recipients == {
-        challenge.creator.email,
-        challenge_admin.email,
-        staff_user.email,
-    }
-
-    challenge_admin_email = [
-        m for m in mail.outbox if challenge_admin.email in m.to
-    ]
-    assert (
-        challenge_admin_email[0].subject
-        == "[testserver] [test] over 70% Budget Consumed Alert"
-    )
-    assert (
-        "We would like to inform you that more than 70% of the compute budget for "
-        "the test challenge has been used." in challenge_admin_email[0].body
-    )
-
-    mail.outbox.clear()
-    evaluation = EvaluationFactory(
-        submission__phase=phase,
-        time_limit=60,
-    )
-    evaluation.utilization.invoice = invoice
-    evaluation.utilization.compute_cost_euro_millicents = 100000
-    evaluation.utilization.save()
-    update_challenge_compute_costs()
-
-    # Next budget alert threshold not exceeded
-    assert len(mail.outbox) == 0
-
-    evaluation = EvaluationFactory(
-        submission__phase=phase,
-        time_limit=60,
-    )
-    evaluation.utilization.invoice = invoice
-    evaluation.utilization.compute_cost_euro_millicents = 1
-    evaluation.utilization.save()
-    update_challenge_compute_costs()
-
-    # Next budget alert threshold exceeded
-    assert len(mail.outbox) != 0
-    assert (
-        mail.outbox[0].subject
-        == "[testserver] [test] over 90% Budget Consumed Alert"
-    )
-
-
-@pytest.mark.django_db
-def test_challenge_budget_alert_two_thresholds_one_email(settings):
-    challenge = ChallengeFactory(short_name="test")
-    assert challenge.percent_budget_consumed_warning_thresholds == [
-        70,
-        90,
-        100,
-    ]
-    challenge_admin = UserFactory()
-    challenge.add_admin(challenge_admin)
-    staff_user = UserFactory(is_staff=True)
-    settings.MANAGERS = [(staff_user.last_name, staff_user.email)]
-    invoice = InvoiceFactory(
-        challenge=challenge,
-        support_costs_euros=0,
-        compute_costs_euros=10,
-        storage_costs_euros=0,
-        payment_status=PaymentStatusChoices.PAID,
-    )
-    phase = PhaseFactory(challenge=challenge)
-    evaluation = EvaluationFactory(
-        submission__phase=phase,
-        time_limit=60,
-    )
-    evaluation.utilization.invoice = invoice
-    evaluation.utilization.compute_cost_euro_millicents = 950000
-    evaluation.utilization.save()
-    update_challenge_compute_costs()
-
-    # Two budget alert thresholds exceeded, alert only sent for last one.
-    assert len(mail.outbox) == 3
-    recipients = {r for m in mail.outbox for r in m.to}
-    assert recipients == {
-        challenge.creator.email,
-        challenge_admin.email,
-        staff_user.email,
-    }
-    assert (
-        mail.outbox[0].subject
-        == "[testserver] [test] over 90% Budget Consumed Alert"
-    )
-
-
-@pytest.mark.django_db
-def test_challenge_budget_alert_no_budget():
-    challenge = ChallengeFactory()
-    phase = PhaseFactory(challenge=challenge)
-    invoice = InvoiceFactory(challenge=challenge, compute_costs_euros=0)
-    evaluation = EvaluationFactory(
-        submission__phase=phase,
-        time_limit=60,
-    )
-    evaluation.utilization.invoice = invoice
-    evaluation.utilization.compute_cost_euro_millicents = 1
-    evaluation.utilization.save()
-    assert len(mail.outbox) == 0
-    update_challenge_compute_costs()
-    assert len(mail.outbox) != 0
-    assert "Budget Consumed Alert" in mail.outbox[0].subject
 
 
 _fixed_now = datetime(2025, 1, 29, 11, 0, 0, tzinfo=ZoneInfo("UTC"))
